Fix MS Clang warnings - use size_t

+++ include/dsa/List.h
@@ -797,7 +797,7 @@ namespace dsa
          * @retval Node* if index is valid
          * @retval nullptr if invalid index
          */
-        Node* get(int index) const;
+        Node* get(size_t index) const;
 
         /**
          * @brief Function sets value of specifed Node of List
@@ -808,7 +808,7 @@ namespace dsa
          * @retval true if value of Node was overwritten
          * @retval false if invalid index
          */
-        bool set(int index, T value);
+        bool set(size_t index, T value);
 
     private:
 
@@ -986,7 +986,7 @@ namespace dsa
     template<typename T>
     List<T>::List(const List<T>& other)
     {
-        for (int i = 0; i < other.size(); i++)
+        for (size_t i = 0; i < other.size(); i++)
         {
             push_back(other.get(i)->value());
         }
@@ -1002,7 +1002,7 @@ namespace dsa
                 pop_front();
             }
 
-            for (int i = 0; i < other.size(); i++)
+            for (size_t i = 0; i < other.size(); i++)
             {
                 push_back(other.get(i)->value());
             }
@@ -1558,7 +1558,7 @@ namespace dsa
         Node* prev{};
         Node* next{};
 
-        for (int i = 0; i < m_size; i++)
+        for (size_t i = 0; i < m_size; i++)
         {
             next = static_cast<Node*>(temp->m_next);
             temp->m_next = prev;
@@ -1618,7 +1618,7 @@ namespace dsa
     }
 
     template<typename T>
-    typename List<T>::Node* List<T>::get(int index) const
+    typename List<T>::Node* List<T>::get(size_t index) const
     {
         if (index < 0 || index >= m_size)
         {
@@ -1674,7 +1674,7 @@ namespace dsa
     }
 
     template<typename T>
-    bool List<T>::set(int index, T value)
+    bool List<T>::set(size_t index, T value)
     {
         Node* temp = get(index);
         if (temp)
